+++ include/Swapping2D.h
@@ -165,7 +165,7 @@ template<typename real_t> class Swapping2D{
       }
 
 #pragma omp for schedule(guided) nowait
-      for(size_t vtid=0; vtid<_mesh->defOp_scaling_factor*nthreads; ++vtid){
+      for(int vtid=0; vtid<_mesh->defOp_scaling_factor*nthreads; ++vtid){
         _mesh->commit_swapping_propagation(marked_edges, vtid);
       }
 
@@ -300,7 +300,7 @@ template<typename real_t> class Swapping2D{
             }
 
 #pragma omp for schedule(guided)
-            for(size_t vtid=0; vtid<_mesh->defOp_scaling_factor*nthreads; ++vtid){
+            for(int vtid=0; vtid<_mesh->defOp_scaling_factor*nthreads; ++vtid){
               _mesh->commit_deferred(vtid);
               _mesh->commit_swapping_propagation(marked_edges, vtid);
               _mesh->commit_colour_reset(node_colour, vtid);
